@@ -1,6 +1,6 @@
 {
   "name": "helloworld",
-  "version": "2.0.16-beta.56",
+  "version": "2.0.17-beta.3",
   "description": "Hello world for metadata.js",
   "repository": {
     "type": "git",
@@ -27,50 +27,50 @@
   },
   "homepage": "https://hello.oknosoft.ru/",
   "dependencies": {
-    "metadata-abstract-ui": "2.0.16-beta.56",
-    "metadata-core": "2.0.16-beta.56",
-    "metadata-external": "2.0.16-beta.56",
-    "metadata-pouchdb": "2.0.16-beta.56",
-    "metadata-react": "2.0.16-beta.56",
-    "metadata-redux": "2.0.16-beta.56"
+    "metadata-abstract-ui": "2.0.17-beta.3",
+    "metadata-core": "2.0.17-beta.3",
+    "metadata-external": "2.0.17-beta.3",
+    "metadata-pouchdb": "2.0.17-beta.3",
+    "metadata-react": "2.0.17-beta.3",
+    "metadata-redux": "2.0.17-beta.3"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.2.2",
-    "babel-jest": "^22.4.3",
-    "babel-loader": "^7.1.4",
-    "babel-preset-react-app": "^3.1.1",
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^8.2.3",
+    "babel-jest": "^23.0.1",
+    "babel-loader": "^7.1.5",
+    "babel-preset-react-app": "^3.1.2",
     "babel-runtime": "^6.26.0",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "^2.3.2",
+    "case-sensitive-paths-webpack-plugin": "2.1.2",
+    "chalk": "^2.4.1",
     "css-loader": "^0.28.11",
-    "dotenv": "^5.0.1",
+    "dotenv": "^6.0.0",
     "eslint": "^4.19.1",
     "eslint-config-react-app": "^2.1.0",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-babel": "^4.1.2",
-    "eslint-plugin-flowtype": "^2.46.1",
-    "eslint-plugin-import": "^2.10.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
+    "eslint-loader": "^2.0",
+    "eslint-plugin-babel": "^5.1",
+    "eslint-plugin-flowtype": "^2.49.3",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jsx-a11y": "^6.1.0",
+    "eslint-plugin-react": "^7.9.1",
     "extract-text-webpack-plugin": "3.0.2",
     "file-loader": "^1.1.11",
-    "fs-extra": "^5.0.0",
-    "html-webpack-plugin": "^2.30.1",
-    "jest": "^22.4.3",
-    "md5-file": "^3.2.3",
-    "postcss-flexbugs-fixes": "^3.3.0",
-    "postcss-loader": "^2.1.3",
+    "fs-extra": "^6.0.0",
+    "html-webpack-plugin": "^3.2.0",
+    "jest": "^23.1.0",
+    "md5-file": "^4.0.0",
+    "postcss-flexbugs-fixes": "^3.3.1",
+    "postcss-loader": "^2.1.5",
     "raf": "3.4.0",
     "raw-loader": "^0.5.1",
-    "react": "^16.3.1",
+    "react": "^16.4.1",
     "react-dev-utils": "^5.0.1",
-    "react-dom": "^16.3.1",
+    "react-dom": "^16.4.1",
     "rimraf": "^2.6",
-    "style-loader": "^0.19.1",
+    "style-loader": "^0.21.0",
     "sw-precache-webpack-plugin": "^0.11.5",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.11.0",
+    "url-loader": "^1.0.1",
+    "webpack": "^3.12.0",
     "webpack-dev-server": "^2.11.2",
     "webpack-manifest-plugin": "^1.3.2"
   },
