@@ -1,12 +1,12 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import {Switch, Route} from 'react-router';
-import Typography from 'material-ui/Typography';
-import AppBar from 'material-ui/AppBar';
-import Toolbar from 'material-ui/Toolbar';
-import IconButton from 'material-ui/IconButton';
+import Typography from '@material-ui/core/Typography';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
-//import Tooltip from 'material-ui/Tooltip';
+//import Tooltip from '@material-ui/core/Tooltip';
 import Snack from 'metadata-react/App/Snack';       // сообщения в верхней части страницы (например, обновить после первого запуска)
 import Alert from 'metadata-react/App/Alert';       // диалог сообщения пользователю
 import Confirm from 'metadata-react/App/Confirm';   // диалог вопросов пользователю (да, нет)
@@ -18,7 +18,7 @@ import HeaderButtons from 'metadata-react/Header/HeaderButtons';
 import DumbScreen from '../DumbScreen';             // заставка "загрузка занных"
 import DataRoute from '../DataRoute';               // вложенный маршрутизатор страниц с данными
 import MarkdownRoute from '../MarkdownRoute';       // вложенный маршрутизатор страниц с Markdown, 404 живёт внутри Route
-import HomeView from '../../pages/Home';            // домашняя страница
+import HomeView from '../Home';            // домашняя страница
 import MetaTreePage from '../MetaTreePage';         // дерево метаданных
 import Settings from '../Settings';                 // страница настроек приложения
 
@@ -27,7 +27,7 @@ import withStyles from './styles';
 import withWindowSize from 'metadata-react/WindowSize';
 import compose from 'recompose/compose';
 
-import items, {item_props} from '../../pages';      // массив элементов меню и метод для вычисления need_meta, need_user по location.pathname
+import items, {item_props} from './menu_items';      // массив элементов меню и метод для вычисления need_meta, need_user по location.pathname
 
 // основной layout
 class AppView extends Component {
@@ -45,7 +45,6 @@ class AppView extends Component {
   }
 
   shouldComponentUpdate(props, {need_user, need_meta}) {
-    const {meta_loaded, user, offline} = props;
     const iprops = item_props();
     let res = true;
 
@@ -56,12 +55,6 @@ class AppView extends Component {
 
     if(need_meta != !!iprops.need_meta) {
       this.setState({need_meta: !!iprops.need_meta});
-      res = false;
-    }
-
-    // если есть сохранённый пароль и online, пытаемся авторизоваться
-    if(meta_loaded && !user.logged_in && user.has_login && !user.try_log_in && !offline) {
-      props.handleLogin();
       res = false;
     }
 
@@ -123,14 +116,22 @@ class AppView extends Component {
         dstyle.marginLeft = dx;
       }
 
-      if(meta_loaded && state.need_user && ((!user.try_log_in && !user.logged_in) || (couch_direct && offline))) {
-        return <NeedAuth
-          key="auth"
-          handleNavigate={handleNavigate}
-          handleIfaceState={props.handleIfaceState}
-          title={title}
-          offline={couch_direct && offline}
-        />;
+      let need_auth = meta_loaded && state.need_user && ((!user.try_log_in && !user.logged_in) || (couch_direct && offline));
+      if(need_auth && !couch_direct && $p.current_user && $p.current_user.name == user.name) {
+        need_auth = false;
+      }
+
+      if(need_auth) {
+        return (
+          <div style={dstyle}>
+            <NeedAuth
+              handleNavigate={handleNavigate}
+              handleIfaceState={props.handleIfaceState}
+              title={title}
+              offline={couch_direct && offline}
+            />
+          </div>
+        );
       }
 
       if(!location.pathname.match(/\/login$/) && ((state.need_meta && !meta_loaded) || (state.need_user && !props.complete_loaded))) {
@@ -178,18 +179,17 @@ class AppView extends Component {
     return [
       // основной layout
       <div key="content" className={classes.root}>
-        <AppBar className={appBarClassName}>
+        <AppBar className={appBarClassName} color="default">
           <Toolbar disableGutters>
             <IconButton
-              color="inherit"
               aria-label="open drawer"
               onClick={this.handleDrawerToggle}
               className={navIconClassName}
             >
-              <MenuIcon/>
+              <MenuIcon color="inherit"/>
             </IconButton>
 
-            <Typography className={classes.title} variant="title" color="inherit" noWrap>{title || 'Metadata'}</Typography>
+            <Typography className={classes.title} variant="title" color="textSecondary" noWrap>{title || 'Metadata'}</Typography>
 
             <HeaderButtons
               sync_started={sync_started}
@@ -197,6 +197,7 @@ class AppView extends Component {
               offline={offline}
               user={user}
               handleNavigate={handleNavigate}
+              barColor="default"
             />
 
           </Toolbar>
