@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Paper from 'material-ui/Paper';
-import Typography from 'material-ui/Typography';
+import Paper from '@material-ui/core/Paper';
+import Typography from '@material-ui/core/Typography';
 import Helmet from 'react-helmet';
 import CnnSettings from 'metadata-react/FrmLogin/CnnSettings';
 import compose from 'recompose/compose';
